Split up trakt/tmdb components for platforms
