+++ api/trakt/src/commonMain/kotlin/app/tivi/trakt/TraktComponent.kt
@@ -24,58 +24,13 @@ import app.moviebase.trakt.api.TraktSeasonsApi
 import app.moviebase.trakt.api.TraktShowsApi
 import app.moviebase.trakt.api.TraktSyncApi
 import app.moviebase.trakt.api.TraktUsersApi
-import app.tivi.data.traktauth.RefreshTraktTokensInteractor
-import app.tivi.data.traktauth.TraktOAuthInfo
-import app.tivi.data.traktauth.store.AuthStore
-import app.tivi.inject.ApplicationScope
-import io.ktor.client.engine.okhttp.OkHttp
-import io.ktor.client.plugins.auth.Auth
-import io.ktor.client.plugins.auth.providers.BearerTokens
-import io.ktor.client.plugins.auth.providers.bearer
 import me.tatarka.inject.annotations.Provides
-import okhttp3.OkHttpClient
 
-interface TraktComponent {
+interface TraktComponent : TraktCommonComponent, TraktPlatformComponent
 
-    @ApplicationScope
-    @Provides
-    fun provideTrakt(
-        client: OkHttpClient,
-        authStore: AuthStore,
-        oauthInfo: TraktOAuthInfo,
-        refreshTokens: Lazy<RefreshTraktTokensInteractor>,
-    ): Trakt = Trakt {
-        traktApiKey = oauthInfo.clientId
-        maxRetriesOnException = 3
+expect interface TraktPlatformComponent
 
-        httpClient(OkHttp) {
-            // Probably want to move to using Ktor's caching, timeouts, etc eventually
-            engine {
-                preconfigured = client
-            }
-
-            install(Auth) {
-                bearer {
-                    loadTokens {
-                        authStore.get()?.let {
-                            BearerTokens(it.accessToken, it.refreshToken)
-                        }
-                    }
-
-                    refreshTokens {
-                        refreshTokens.value.invoke()?.let {
-                            BearerTokens(it.accessToken, it.refreshToken)
-                        }
-                    }
-
-                    sendWithoutRequest { request ->
-                        request.url.host == "api.trakt.tv"
-                    }
-                }
-            }
-        }
-    }
-
+interface TraktCommonComponent {
     @Provides
     fun provideTraktUsersService(trakt: Trakt): TraktUsersApi = trakt.users
